Added HasMeta trait
